midi.output: make proper typedefs instead of using primitive types
This is needed for two purposes:

1. There was a bug caused by the "PmError = ctypes.c_int" assignment.
It caused any function returning an integer as if it was returning
an error code, and thus we were getting Exceptions for integers.
Now the bug is fixed because PmError is a subclass of ctypes.c_int,
so we can safely do: "if restype == PmError: ...".

2. Type checking.
We should be careful with calling the C code, because it crashes the
application if we mess up with function arguments.
So those additional typedefs should help us to avoid such crashes
by enforcing some additional type checks.